Перенесет рендеринг доски и карточек в контроллер
