+++ src/components/show-more-button.js
@@ -8,4 +8,8 @@ export default class SortingMenu extends AbstractComponent {
   getTemplate() {
     return createShowMoreButtonTemplate();
   }
+
+  setOnClick(handler) {
+    this.getElement().addEventListener(`click`, handler);
+  }
 }
