--- src/controllers/board.js
@@ -1,4 +1,0 @@
-
-export default class BoardController {
-
-}
